+++ src/Ssh.php
@@ -2,19 +2,25 @@
 
 namespace Spatie\Ssh;
 
+use Closure;
+use Exception;
 use Symfony\Component\Process\Process;
 
 class Ssh
 {
-    private string $user;
+    protected string $user;
 
-    private string $host;
+    protected string $host;
 
-    private string $pathToPrivateKey = '';
+    protected string $pathToPrivateKey = '';
 
-    private ?int $port;
+    protected ?int $port;
 
-    private bool $enableStrictHostChecking = true;
+    protected bool $enableStrictHostChecking = true;
+
+    protected Closure $processConfigurationClosure;
+
+    protected Closure $onOutput;
 
     public function __construct(string $user, string $host, int $port = null)
     {
@@ -23,6 +29,10 @@ class Ssh
         $this->host = $host;
 
         $this->port = $port;
+
+        $this->processConfigurationClosure = fn(Process $process) => null;
+
+        $this->onOutput = fn($type, $line) => null;
     }
 
     public static function create(...$args): self
@@ -39,7 +49,24 @@ class Ssh
 
     public function usePort(int $port): self
     {
+        if ($port < 0) {
+            throw new Exception('Port must be a positive integer.');
+        }
         $this->port = $port;
+
+        return $this;
+    }
+
+    public function configureProcess(Closure $processConfigurationClosure): self
+    {
+        $this->processConfigurationClosure = $processConfigurationClosure;
+
+        return $this;
+    }
+
+    public function onOutput(Closure $onOutput): self
+    {
+        $this->onOutput = $onOutput;
 
         return $this;
     }
@@ -63,19 +90,19 @@ class Ssh
      *
      * @return string
      */
-    public function getSshCommand($command): string
+    public function getExecuteCommand($command): string
     {
         $commands = $this->wrapArray($command);
 
-        $extraOptions = $this->getExtraOptions();
+        $extraOptions = $this->getExtraSshOptions();
 
         $commandString = implode(PHP_EOL, $commands);
 
         $delimiter = 'EOF-SPATIE-SSH';
 
-        $target = "{$this->user}@{$this->host}";
+        $target = $this->getTarget();
 
-        return "ssh {$extraOptions} $target 'bash -se' << \\$delimiter".PHP_EOL
+        return "ssh {$extraOptions} {$target} 'bash -se' << \\$delimiter".PHP_EOL
             .$commandString.PHP_EOL
             .$delimiter;
     }
@@ -87,23 +114,72 @@ class Ssh
      */
     public function execute($command): Process
     {
-        $sshCommand = $this->getSshCommand($command);
+        $sshCommand = $this->getExecuteCommand($command);
 
-        $process = Process::fromShellCommandline($sshCommand);
-
-        $process->setTimeout(0);
-
-        $process->run();
-
-        return $process;
+        return $this->run($sshCommand);
     }
 
-    protected function wrapArray($arrayOrString): array
+    /**
+     * @param string|array $command
+     *
+     * @return \Symfony\Component\Process\Process
+     */
+    public function executeAsync($command): Process
     {
-        return (array) $arrayOrString;
+        $sshCommand = $this->getExecuteCommand($command);
+
+        return $this->run($sshCommand, 'start');
     }
 
-    protected function getExtraOptions(): string
+    public function getDownloadCommand(string $sourcePath, string $destinationPath): string
+    {
+        return "scp {$this->getExtraScpOptions()} {$this->getTarget()}:$sourcePath $destinationPath";
+    }
+
+    public function download(string $sourcePath, string $destinationPath): Process
+    {
+        $downloadCommand = $this->getDownloadCommand($sourcePath, $destinationPath);
+
+        return $this->run($downloadCommand);
+    }
+
+    public function getUploadCommand(string $sourcePath, string $destinationPath): string
+    {
+        return "scp {$this->getExtraScpOptions()} $sourcePath {$this->getTarget()}:$destinationPath";
+    }
+
+    public function upload(string $sourcePath, string $destinationPath): Process
+    {
+        $uploadCommand = $this->getUploadCommand($sourcePath, $destinationPath);
+
+        return $this->run($uploadCommand);
+    }
+
+    protected function getExtraSshOptions(): string
+    {
+        $extraOptions = $this->getExtraOptions();
+
+        if (! is_null($this->port)) {
+            $extraOptions[] = "-p {$this->port}";
+        }
+
+        return implode(' ', $extraOptions);
+    }
+
+    protected function getExtraScpOptions(): string
+    {
+        $extraOptions = $this->getExtraOptions();
+
+        $extraOptions[] = '-r';
+
+        if (! is_null($this->port)) {
+            $extraOptions[] = "-P {$this->port}";
+        }
+
+        return implode(' ', $extraOptions);
+    }
+
+    private function getExtraOptions(): array
     {
         $extraOptions = [];
 
@@ -111,15 +187,34 @@ class Ssh
             $extraOptions[] = "-i {$this->pathToPrivateKey}";
         }
 
-        if ($this->port) {
-            $extraOptions[] = "-p {$this->port}";
-        }
-
         if (! $this->enableStrictHostChecking) {
             $extraOptions[] = '-o StrictHostKeyChecking=no';
             $extraOptions[] = '-o UserKnownHostsFile=/dev/null';
         }
 
-        return implode(' ', $extraOptions);
+        return $extraOptions;
+    }
+
+    protected function wrapArray($arrayOrString): array
+    {
+        return (array) $arrayOrString;
+    }
+
+    protected function run(string $command, string $method = 'run'): Process
+    {
+        $process = Process::fromShellCommandline($command);
+
+        $process->setTimeout(0);
+
+        ($this->processConfigurationClosure)($process);
+
+        $process->{$method}($this->onOutput);
+
+        return $process;
+    }
+
+    protected function getTarget(): string
+    {
+        return "{$this->user}@{$this->host}";
     }
 }