+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to `ssh` will be documented in this file
 
+## 1.4.0 - 2020-04-10
+
+- add `executeAsync`
+
+## 1.3.0 - 2020-02-14
+
+- add `onOutput`
+
+## 1.2.0 - 2020-02-14
+
+- add `upload`, `download` and `configureProcess` methods
+
+## 1.1.2 - 2020-02-13
+
+- allow 0 as a valid port number
+
 ## 1.1.1 - 2020-02-10
 
 - for security reasons, the host key checking is now enabled by default